+++ userbot/plugins/nsfw.py
@@ -2,44 +2,51 @@
 # Don't edit credits Madafaka
 """
 This module can search images in danbooru and send in to the chat!
-
 ──「 **Danbooru Search** 」──
 """
-
+ 
 import os
 import urllib
-from asyncio import sleep
-
+ 
 import requests
-
+ 
 from userbot import catub
-
-from ..core.managers import edit_delete, edit_or_reply
-
+ 
+from ..helpers.functions import age_verification
+from . import edit_delete, edit_or_reply, reply_id
+ 
 plugin_category = "fun"
-
-
+ 
+ 
 @catub.cat_cmd(
     pattern="ani(mu|nsfw) ?(.*)",
-    command=("animu|aninsfw", plugin_category),
+    command=("ani", plugin_category),
     info={
         "header": "Contains NSFW 🔞.\nTo search images in danbooru!",
-        "usage": "{tr}animu <query>\n{tr}aninsfw <nsfw query>",
-        "examples": "{tr}animu One punch man",
+        "usage": [
+            "{tr}animu <query>",
+            "{tr}aninsfw <nsfw query>",
+        ],
+        "examples": [
+            "{tr}animu naruto",
+            "{tr}aninsfw boku no pico",
+        ],
     },
 )
 async def danbooru(message):
+    "Get anime charecter pic or nsfw"
+    reply_to = await reply_id(e)
+    if await age_verification(e, reply_to):
+        return
     await edit_or_reply(message, "`Processing…`")
-
     rating = "Explicit" if "nsfw" in message.pattern_match.group(1) else "Safe"
     search_query = message.pattern_match.group(2)
-
     params = {
         "limit": 1,
         "random": "true",
         "tags": f"Rating:{rating} {search_query}".strip(),
     }
-
+ 
     with requests.get(
         "http://danbooru.donmai.us/posts.json", params=params
     ) as response:
@@ -52,17 +59,17 @@ async def danbooru(message):
                 4,
             )
             return
-
+ 
     if not response:
         await edit_delete(message, f"`No results for query:` __{search_query}__", 4)
         return
-
+ 
     valid_urls = [
         response[0][url]
         for url in ["file_url", "large_file_url", "source"]
         if url in response[0].keys()
     ]
-
+ 
     if not valid_urls:
         await edit_delete(
             message, f"`Failed to find URLs for query:` __{search_query}__", 4
@@ -70,7 +77,9 @@ async def danbooru(message):
         return
     for image_url in valid_urls:
         try:
-            await message.client.send_file(message.chat_id, image_url)
+            await message.client.send_file(
+                message.chat_id, image_url, reply_to=reply_to
+            )
             await message.delete()
             return
         except Exception as e:
@@ -78,8 +87,8 @@ async def danbooru(message):
     await edit_delete(
         message, f"``Failed to fetch media for query:` __{search_query}__", 4
     )
-
-
+ 
+ 
 @catub.cat_cmd(
     pattern="boobs(?: |$)(.*)",
     command=("boobs", plugin_category),
@@ -90,17 +99,19 @@ async def danbooru(message):
     },
 )
 async def boobs(e):
-    a = await edit_or_reply(e, "`Finding some big boobs...`")
-    await sleep(1)
-    await a.edit("`Sending some big boobs...`")
+    "Search boobs"
+    reply_to = await reply_id(e)
+    if await age_verification(e, reply_to):
+        return
+    a = await edit_or_reply(e, "`Sending boobs...`")
     nsfw = requests.get("http://api.oboobs.ru/noise/1").json()[0]["preview"]
     urllib.request.urlretrieve("http://media.oboobs.ru/{}".format(nsfw), "*.jpg")
     os.rename("*.jpg", "boobs.jpg")
-    await e.client.send_file(e.chat_id, "boobs.jpg")
+    await e.client.send_file(e.chat_id, "boobs.jpg", reply_to=reply_to)
     os.remove("boobs.jpg")
     await a.delete()
-
-
+ 
+ 
 @catub.cat_cmd(
     pattern="butts(?: |$)(.*)",
     command=("butts", plugin_category),
@@ -111,16 +122,18 @@ async def boobs(e):
     },
 )
 async def butts(e):
-    a = await edit_or_reply(e, "`Finding some beautiful butts...`")
-    await sleep(1)
-    await a.edit("`Sending some beautiful butts...`")
+    "Search beautiful butts"
+    reply_to = await reply_id(e)
+    if await age_verification(e, reply_to):
+        return
+    a = await edit_or_reply(e, "`Sending beautiful butts...`")
     nsfw = requests.get("http://api.obutts.ru/butts/10/1/random").json()[0]["preview"]
     urllib.request.urlretrieve("http://media.obutts.ru/{}".format(nsfw), "*.jpg")
     os.rename("*.jpg", "butts.jpg")
-    await e.client.send_file(e.chat_id, "butts.jpg")
+    await e.client.send_file(e.chat_id, "butts.jpg", reply_to=reply_to)
     os.remove("butts.jpg")
     await a.delete()
-
+    
 
 PENIS_TEMPLATE = """
 🍆🍆
